HasImagesAccessor trait
