+++ app/Models/Article.php
@@ -3,11 +3,11 @@
 namespace App\Models;
 
 use App\Traits\Publishable;
+use App\Traits\HasImagesAccessor;
 use App\Traits\HasNavigationHeadings;
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
 use Backpack\CRUD\app\Models\Traits\SpatieTranslatable\HasTranslations;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Http\UploadedFile;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\Sluggable\HasSlug;
@@ -15,7 +15,7 @@ use Spatie\Sluggable\SlugOptions;
 
 class Article extends Model implements HasMedia
 {
-    use CrudTrait, InteractsWithMedia, HasSlug, HasTranslations, Publishable, HasNavigationHeadings;
+    use CrudTrait, InteractsWithMedia, HasSlug, HasTranslations, Publishable, HasNavigationHeadings, HasImagesAccessor;
 
     protected $guarded = ['id'];
     protected $dates = ['published_at'];
@@ -24,21 +24,6 @@ class Article extends Model implements HasMedia
     public function getCoverImageTagAttribute()
     {
         return $this->getFirstMedia()->img();
-    }
-
-    public function getImagesAttribute()
-    {
-        return $this->getMedia();
-    }
-
-    public function setImagesAttribute($uploaded_files)
-    {
-        collect($uploaded_files)
-            ->filter() // Ignore nulls
-            ->each(fn (UploadedFile $file) => $this
-                ->addMedia($file)
-                ->toMediaCollection()
-            );
     }
 
     public function getRouteKeyName()
