+++ app/Traits/HasImagesAccessor.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Traits;
+
+use Illuminate\Http\UploadedFile;
+
+trait HasImagesAccessor
+{
+    public function getImagesAttribute()
+    {
+        return $this->getMedia();
+    }
+
+    public function setImagesAttribute($uploaded_files)
+    {
+        collect($uploaded_files)
+            ->filter() // Ignore nulls
+            ->each(fn (UploadedFile $file) => $this
+                ->addMedia($file)
+                ->toMediaCollection()
+            );
+    }
+}
